remoe warns

+++ library/newlib/execve.c
@@ -1,11 +1,12 @@
 #include <svc.h>
 #include <errno.h>
+#include <newlib/newlib.h>
 #undef errno
 extern int errno;
 
 int execve(char *path, char **argv, char **env)
 {
-    if (!svc_call(path, argv, env, 0, 0, 0, 0, 0, SYS_EXECVE))
+    if (!svc_call((unsigned long)path, (unsigned long)argv, (unsigned long)env, 0, 0, 0, 0, 0, SYS_EXECVE))
     {
         errno = ENOENT;
         return -1;
